requirements.txt
requests
django
beautifulsoup4
djangorestframework
psycopg2
python-dotenv
fake-useragent